fix #3351 fix #3208
