+++ configuration-client/src/utils/index.tsx
@@ -6,6 +6,13 @@ export const openGitHubNewIssue = () => {
   window.open('https://github.com/UniversalMediaServer/UniversalMediaServer/issues/new','_blank');
 }
 
+export const defaultTooltipSettings = {
+  width: 350,
+  color: 'blue',
+  multiline: true,
+  withArrow: true,
+}
+
 export const aboutApiUrl = '/v1/api/about/';
 export const actionsApiUrl = '/v1/api/actions/';
 export const accountApiUrl = '/v1/api/account/';
