+++ src/main/resources/i18n/messages.properties
@@ -270,6 +270,7 @@ ErrorWhileStartingUms=Error while starting UMS
 ExecutableXTranscodingEngineNotFound=The executable "%1$s" for transcoding engine %2$s could not be found.
 ExternalOutgoingTraffic=External outgoing traffic
 External_abbr=Ext.
+FeedNameNotFound=Feed name not found
 FileExists=File exists
 FileOrder=File order:
 FileSortingNaming=File sorting / naming
